feat: Block group creation with same name

+++ src/storage/group/groupCreate.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from '@react-native-async-storage/async-storage'
+import { AppError } from '@utils/AppError'
 
 import { GROUP_COLLECTION } from '../storageConfig'
 import { getAllGroups } from './getAllGroups'
@@ -6,6 +7,12 @@ import { getAllGroups } from './getAllGroups'
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await getAllGroups()
+
+    const groupAlreadyExists = storedGroups.includes(newGroup)
+
+    if (groupAlreadyExists) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.')
+    }
 
     const storage = JSON.stringify([...storedGroups, newGroup])
     await AsyncStorage.setItem(GROUP_COLLECTION, storage)
